Move Java reflect specific extensions into a separate package
