+++ src/main/java/com/squareup/kotlinpoet/TypeSpec.kt
@@ -16,6 +16,8 @@
 package com.squareup.kotlinpoet
 
 import com.squareup.kotlinpoet.KModifier.PUBLIC
+import com.squareup.kotlinpoet.jvm.asClassName
+import com.squareup.kotlinpoet.jvm.asTypeName
 import java.lang.reflect.Type
 import kotlin.reflect.KClass
